+++ breakpoints/index.js
@@ -1,0 +1,5 @@
+if (process.env.NODE_ENV === 'production') {
+  module.exports = require('./create-breakpoints-api.prod');
+} else {
+  module.exports = require('./create-breakpoints-api.dev');
+}