+++ library/index.js
@@ -1,7 +1,0 @@
-const { get, memoize, createInvariantWithPrefix } = require('./library');
-
-module.exports = {
-  get,
-  memoize,
-  createInvariantWithPrefix,
-};