+++ react-emotion/index.js
@@ -1,5 +1,0 @@
-const { useTheme } = require('@emotion/react');
-
-const { createUseBreakpoint } = require('../hooks/use-breakpoint');
-
-exports.useBreakpoint = createUseBreakpoint({ theme: useTheme });