+++ styled-breakpoints/memoize/index.js
@@ -1,0 +1,1 @@
+exports.memoize = require('./memoize').memoize;