+++ use-media-query/get-matches/index.js
@@ -1,0 +1,1 @@
+exports.getMatches = require('./get-matches').getMatches;