+++ use-media-query/index.d.ts
@@ -1,0 +1,1 @@
+export declare function useMediaQuery(query: string): boolean;