+++ use-media-query/use-media-query/index.js
@@ -1,0 +1,1 @@
+exports.useMediaQuery = require('./use-media-query').useMediaQuery;